🎉 Add initial database schema file
Added the initial database schema file generated by Rails. This file represents the current state of the database structure and will be used as the source of truth for the database schema. The schema is currently empty (version 0) as no migrations have been created yet, but tracking this file in version control will help maintain database consistency across development environments.